+++ internal/pkg/runtime/engine/apptainer/container_linux.go
@@ -14,6 +14,7 @@ import (
 	"fmt"
 	"io"
 	"os"
+	osuser "os/user"
 	"path/filepath"
 	"strconv"
 	"strings"
@@ -89,7 +90,7 @@ type container struct {
 	skippedMount  []string
 	suidFlag      uintptr
 	devSourcePath string
-	imageBind     map[string]string
+	skipCwd       bool
 }
 
 //nolint:maintidx
@@ -107,7 +108,6 @@ func create(ctx context.Context, engine *EngineOperations, rpcOps *client.RPC, p
 		mountInfoPath: fmt.Sprintf("/proc/%d/mountinfo", pid),
 		skippedMount:  make([]string, 0),
 		suidFlag:      syscall.MS_NOSUID,
-		imageBind:     make(map[string]string),
 	}
 
 	cwd := engine.EngineConfig.GetCwd()
@@ -179,6 +179,14 @@ func create(ctx context.Context, engine *EngineOperations, rpcOps *client.RPC, p
 
 	p := &mount.Points{}
 	system := &mount.System{Points: p, Mount: c.mount}
+
+	createCwdDirTag := mount.AuthorizedTag(mount.LayerTag)
+	if c.engine.EngineConfig.GetSessionLayer() == apptainer.UnderlayLayer {
+		createCwdDirTag = mount.PreLayerTag
+	}
+	if err := system.RunBeforeTag(createCwdDirTag, c.createCwdDir); err != nil {
+		return err
+	}
 
 	if err := c.setupSessionLayout(system); err != nil {
 		return err
@@ -1272,9 +1280,9 @@ func (c *container) addImageBindMount(system *mount.System) error {
 				return nil
 			})
 
-			// to honor bind ordering specified by the users
-			// we add bind mount later in addUserbindsMount
-			c.imageBind[destination] = src
+			if err := system.Points.AddBind(mount.UserbindsTag, src, destination, syscall.MS_BIND); err != nil {
+				return fmt.Errorf("while adding data bind %s -> %s: %s", src, destination, err)
+			}
 		}
 	}
 
@@ -1695,9 +1703,8 @@ func (c *container) getHomePaths() (source string, dest string, err error) {
 		dest = filepath.Clean(c.engine.EngineConfig.GetHomeDest())
 		source, err = filepath.Abs(filepath.Clean(c.engine.EngineConfig.GetHomeSource()))
 	} else {
-		pw, err := user.CurrentOriginal()
-		if err == nil {
-			source = pw.Dir
+		source = c.engine.EngineConfig.JSON.UserInfo.Home
+		if source != "" {
 			if c.engine.EngineConfig.GetFakeroot() || os.Getuid() == 0 {
 				// Mount user home directory onto /root for
 				//  any root-mapped namespace
@@ -1849,12 +1856,6 @@ func (c *container) addUserbindsMount(system *mount.System) error {
 		}
 		// data image bind
 		if b.ID() != "" || b.ImageSrc() != "" {
-			source, ok := c.imageBind[b.Destination]
-			if ok {
-				if err := system.Points.AddBind(mount.UserbindsTag, source, b.Destination, syscall.MS_BIND); err != nil {
-					return fmt.Errorf("while adding data bind %s -> %s: %s", source, b.Destination, err)
-				}
-			}
 			continue
 		}
 
@@ -2081,70 +2082,163 @@ func (c *container) isMounted(dest string) bool {
 }
 
 func (c *container) addCwdMount(system *mount.System) error {
-	if c.engine.EngineConfig.GetContain() {
-		sylog.Verbosef("Not mounting current directory: contain was requested")
+	if c.skipCwd {
 		return nil
-	}
-	if !c.engine.EngineConfig.File.UserBindControl {
-		sylog.Warningf("Not mounting current directory: user bind control is disabled by system administrator")
-		return nil
-	}
-	if c.engine.EngineConfig.GetNoCwd() {
-		sylog.Debugf("Skipping current directory mount by user request.")
-		return nil
-	}
-	cwd := c.engine.EngineConfig.GetCwd()
-	if cwd == "" {
-		sylog.Warningf("No current working directory set: skipping mount")
 	}
 
-	current, err := filepath.EvalSymlinks(cwd)
+	cwdHost := filepath.Clean(c.engine.EngineConfig.GetCwd())
+	if cwdHost == "" {
+		sylog.Warningf("No current working directory set: skipping mount")
+		return nil
+	}
+
+	cwdHostResolved, err := filepath.EvalSymlinks(cwdHost)
 	if err != nil {
 		return fmt.Errorf("could not obtain current directory path: %s", err)
 	}
-	sylog.Debugf("Using %s as current working directory", cwd)
 
-	switch current {
-	case "/", "/etc", "/bin", "/mnt", "/usr", "/var", "/opt", "/sbin", "/lib", "/lib64":
-		sylog.Verbosef("Not mounting CWD within operating system directory: %s", current)
-		return nil
-	}
-	if strings.HasPrefix(current, "/sys") || strings.HasPrefix(current, "/proc") || strings.HasPrefix(current, "/dev") {
-		sylog.Verbosef("Not mounting CWD within virtual directory: %s", current)
-		return nil
-	}
+	cwdHostSymlink := cwdHost != cwdHostResolved
 
-	dest := fs.EvalRelative(cwd, c.session.FinalPath())
-	dest = filepath.Join(c.session.FinalPath(), dest)
+	cwdContainerResolved := fs.EvalRelative(cwdHost, c.session.FinalPath())
+	cwdContainerResolved = filepath.Join(c.session.FinalPath(), cwdContainerResolved)
+	cwdContainerSymlink := cwdContainerResolved != cwdHost
 
-	fi, err := c.rpcOps.Stat(dest)
+	fi, err := c.rpcOps.Stat(cwdContainerResolved)
 	if err != nil {
 		if os.IsNotExist(err) {
-			sylog.Verbosef("Not mounting CWD, %s doesn't exist within container", cwd)
+			sylog.Verbosef("Not mounting CWD, %s doesn't exist within container", cwdContainerResolved)
 		}
-		sylog.Verbosef("Not mounting CWD, while getting %s information: %s", cwd, err)
+		sylog.Verbosef("Not mounting CWD, while getting %s information: %s", cwdContainerResolved, err)
 		return nil
 	}
 	cst := fi.Sys().(*syscall.Stat_t)
 
 	var hst syscall.Stat_t
-	if err := syscall.Stat(cwd, &hst); err != nil {
+	if err := syscall.Stat(cwdHost, &hst); err != nil {
 		return err
 	}
+
 	// same ino/dev, the current working directory is available within the container
 	if hst.Dev == cst.Dev && hst.Ino == cst.Ino {
-		sylog.Verbosef("%s found within container", cwd)
+		sylog.Verbosef("%s found within container", cwdHost)
 		return nil
-	} else if c.isMounted(dest) {
-		sylog.Verbosef("Not mounting CWD (already mounted in container): %s", cwd)
+	} else if c.isMounted(cwdContainerResolved) {
+		sylog.Verbosef("Not mounting CWD (already mounted in container): %s", cwdHost)
+		return nil
+	} else if cwdHostSymlink && cwdContainerSymlink && cwdContainerResolved != cwdHostResolved {
+		// symlink case when both destination exists on host and in container but are differents
+		sylog.Verbosef("Not mounting CWD, detected symlinks with different destination between host/container")
 		return nil
 	}
 
 	flags := uintptr(syscall.MS_BIND | c.suidFlag | syscall.MS_NODEV | syscall.MS_REC)
-	if err := system.Points.AddBind(mount.CwdTag, cwd, cwd, flags); err != nil {
-		return fmt.Errorf("could not bind cwd directory %s into container: %s", cwd, err)
+	if err := system.Points.AddBind(mount.CwdTag, cwdHost, cwdHost, flags); err != nil {
+		if errors.Is(err, mount.ErrMountExists) {
+			return nil
+		}
+		return fmt.Errorf("could not bind cwd directory %s into container: %s", cwdHost, err)
 	}
-	return system.Points.AddRemount(mount.CwdTag, cwd, flags)
+	return system.Points.AddRemount(mount.CwdTag, cwdHost, flags)
+}
+
+func (c *container) createCwdDir(system *mount.System) error {
+	if c.engine.EngineConfig.GetContain() {
+		c.skipCwd = true
+		sylog.Verbosef("Not mounting current directory: contain was requested")
+		return nil
+	}
+	if !c.engine.EngineConfig.File.UserBindControl {
+		c.skipCwd = true
+		sylog.Warningf("Not mounting current directory: user bind control is disabled by system administrator")
+		return nil
+	}
+	if c.engine.EngineConfig.GetNoCwd() {
+		c.skipCwd = true
+		sylog.Debugf("Skipping current directory mount by user request.")
+		return nil
+	}
+
+	cwdHost := filepath.Clean(c.engine.EngineConfig.GetCwd())
+	if cwdHost == "" || cwdHost == "/" {
+		c.skipCwd = true
+		sylog.Warningf("No current working directory set: skipping mount")
+		return nil
+	} else if cwdHost[0] != '/' {
+		return fmt.Errorf("current working directory %s is not an absolute path", cwdHost)
+	}
+
+	cwdHostResolved, err := filepath.EvalSymlinks(cwdHost)
+	if err != nil {
+		return fmt.Errorf("could not obtain current directory path: %s", err)
+	}
+	sylog.Debugf("Using %s as current working directory", cwdHost)
+
+	cwdPaths := []string{cwdHost}
+	cwdHostSymlink := cwdHost != cwdHostResolved
+
+	// handle new symlink /bin -> usr/bin, /sbin -> usr/sbin ...
+	if cwdHostSymlink {
+		cwdPaths = append(cwdPaths, cwdHostResolved)
+	}
+
+	for _, cwdPath := range cwdPaths {
+		switch cwdPath {
+		case "/", "/etc", "/bin", "/mnt", "/usr", "/var", "/opt", "/sbin", "/lib", "/lib64":
+			sylog.Verbosef("Not mounting CWD within operating system directory: %s", cwdPath)
+			c.skipCwd = true
+			return nil
+		}
+		if strings.HasPrefix(cwdPath, "/sys") || strings.HasPrefix(cwdPath, "/proc") || strings.HasPrefix(cwdPath, "/dev") {
+			sylog.Verbosef("Not mounting CWD within virtual directory: %s", cwdPath)
+			c.skipCwd = true
+			return nil
+		}
+	}
+
+	pathComponents := strings.Split(cwdHost, string(os.PathSeparator))
+	existingPath := false
+	path := ""
+
+	// check if first or last path component exists in container either
+	// in container image or via user/home binds
+	for i, pathComponent := range pathComponents[1:] {
+		if i == 0 {
+			path = filepath.Join(string(os.PathSeparator), pathComponent)
+		} else {
+			path = filepath.Join(path, pathComponent)
+		}
+		_, err := c.session.GetOverridePath(path)
+		existingPath = err == nil
+		if err != nil {
+			pathContainerResolved := fs.EvalRelative(path, c.session.RootFsPath())
+			if pathContainerResolved != path {
+				return nil
+			}
+			info, err := c.rpcOps.Lstat(filepath.Join(c.session.RootFsPath(), pathContainerResolved))
+			if err != nil {
+				existingPath = false
+			} else if !info.IsDir() {
+				existingPath = true
+			}
+		}
+	}
+
+	if !existingPath && c.session.Layer != nil {
+		if c.engine.EngineConfig.GetSessionLayer() == apptainer.UnderlayLayer {
+			flags := uintptr(syscall.MS_BIND | c.suidFlag | syscall.MS_NODEV | syscall.MS_REC)
+			if err := system.Points.AddBind(mount.CwdTag, cwdHost, cwdHost, flags); err != nil {
+				return fmt.Errorf("could not bind cwd directory %s into container: %s", cwdHost, err)
+			}
+			return system.Points.AddRemount(mount.CwdTag, cwdHost, flags)
+		}
+		sessionPath := filepath.Join(c.session.Layer.Dir(), cwdHost)
+		// ignore error and let addCwdMount failing properly
+		if err := c.session.AddDir(sessionPath); err != nil {
+			sylog.Warningf("Not creating container current working directory: %s", err)
+		}
+	}
+
+	return nil
 }
 
 func (c *container) addLibsMount(system *mount.System) error {
@@ -2269,7 +2363,7 @@ func (c *container) addIdentityMount(system *mount.System) error {
 		if err != nil {
 			sylog.Warningf("%s", err)
 		} else {
-			content, err := files.Passwd(passwd, home, uid)
+			content, err := files.Passwd(passwd, home, uid, c)
 			if err != nil {
 				sylog.Warningf("%s", err)
 			} else {
@@ -2292,7 +2386,7 @@ func (c *container) addIdentityMount(system *mount.System) error {
 
 	if c.engine.EngineConfig.File.ConfigGroup {
 		group := filepath.Join(rootfs, "/etc/group")
-		content, err := files.Group(group, uid, c.engine.EngineConfig.GetTargetGID())
+		content, err := files.Group(group, uid, c.engine.EngineConfig.GetTargetGID(), c)
 		if err != nil {
 			sylog.Warningf("%s", err)
 		} else {
@@ -2825,4 +2919,37 @@ func gocryptfsMount(params *image.MountParams, mfunc image.MountFunc) error {
 		return fmt.Errorf("could not locate the decrypted squashfs file, previous gocryptfs mount failed")
 	}
 	return imageDriver.Mount(params, mfunc)
+}
+
+func (c *container) GetPwUID(uid uint32) (*user.User, error) {
+	if c.engine.EngineConfig.JSON.UserInfo.Username == "" {
+		return nil, osuser.UnknownUserIdError(uid)
+	}
+	return &user.User{
+		Name:  c.engine.EngineConfig.JSON.UserInfo.Username,
+		UID:   uint32(c.engine.EngineConfig.JSON.UserInfo.UID),
+		GID:   uint32(c.engine.EngineConfig.JSON.UserInfo.GID),
+		Gecos: c.engine.EngineConfig.JSON.UserInfo.Gecos,
+		Dir:   c.engine.EngineConfig.JSON.UserInfo.Home,
+		Shell: c.engine.EngineConfig.JSON.UserInfo.Shell,
+	}, nil
+}
+
+func (c *container) GetGrGID(gid uint32) (*user.Group, error) {
+	name, ok := c.engine.EngineConfig.JSON.UserInfo.Groups[int(gid)]
+	if ok {
+		return &user.Group{
+			Name: name,
+			GID:  gid,
+		}, nil
+	}
+	return nil, osuser.UnknownGroupIdError(fmt.Sprintf("%d", gid))
+}
+
+func (c *container) Getgroups() ([]int, error) {
+	gids := make([]int, 0, len(c.engine.EngineConfig.JSON.UserInfo.Groups))
+	for gid := range c.engine.EngineConfig.JSON.UserInfo.Groups {
+		gids = append(gids, gid)
+	}
+	return gids, nil
 }